refactor: add IndexPage
